--- src/AnalyzeType.css
@@ -1,67 +1,0 @@
-/*#presentation-back{
-background: url(/public/background-img.png);
-}*/
-#presentation-type{
-   margin-left: 20%;
-   margin-right:20%;
-   margin-top: 2%;
-}
-
-#info-type-one{
-  margin-left: 20%;
-  margin-right: 20%;
-  width: 60%;
-  display: inline-block;
-  text-align: center;
- /*background-color: rgba(249, 184, 184, 0.25);*/
-  background-color: rgba(237, 242, 95, 0.1);
-  border-radius: 8%;
-  color: black;
-  text-decoration: none;
-}
-
-#type-title{
-   font-size: 25pt;
-   margin-top: 10%;
-}
-
-#type-explain{
-   font-size: 12pt;
-   white-space: pre-line;
-   margin-bottom: 10%;
-}
-
-/*#info-type-two{
-  margin-right: 2%;
-  width: 32%;
-  display: inline-block;
-  text-align: center;
-  background-color: rgba(240, 124, 66, 0.17);
-  border-radius: 8%;
-  color: black;
-  text-decoration: none;
-}*/
-
-#info-type-two{
-  margin-left: 20%;
-  margin-right: 20%;
-  margin-top: 3%;
-  width: 60%;
-  display: inline-block;
-  text-align: center;
-  background-color: rgba(95, 181, 242, 0.1);
-  border-radius: 8%;
-  color: black;
-  text-decoration: none;
-}
-
-/*#info-type-four{
-  margin-right: 2%;
-  width: 32%;
-  display: inline-block;
-  text-align: center;
-  border-radius: 8%;
-  color: black;
-  text-decoration: none;
-}
-*/
