+++ src/App.js
@@ -1,23 +1,28 @@
 //import logo from './logo.svg';
 import './App.css';
-import Navbar from './Navbar.js';
-import AnalyzeType from './AnalyzeType.js';
+import Navbar from './components/Navbar.js';
+import Footer from './components/Footer.js';
+import AnalyzeType from './pages/AnalyzeType.js';
+import Home from './pages/Home.js';
+import SignInPage from './pages/SignIn';
+import SignUpPage from './pages/SignUp';
 import {BrowserRouter, Route, Routes, Navigate, Outlet} from 'react-router-dom';
-import SignInPage from './SignIn';
-import SignUpPage from './SignUp';
+
 
 function App() {
   return (
-  <>
+  <div className="App">
   <BrowserRouter>
     <Navbar />
     <Routes>
+      <Route path='/' element={<Home/>}/>
       <Route path="/analyzetype" element={<AnalyzeType />} />
       <Route path="/signin" element={<SignInPage />} />
       <Route path="/signup" element={<SignUpPage />} />
     </Routes>
+    <Footer/>
   </BrowserRouter>
-  </>
+  </div>
   );
 }
