+++ src/pages/SignUp.js
@@ -1,4 +1,4 @@
-import './Sign.css';
+import '../styles/Sign.css';
 import { Link } from 'react-router-dom';
 
 function SignUpPage(){
