+++ src/styles/Footer.css
@@ -1,0 +1,21 @@
+.footer{
+    width:100%;
+    height:200px;
+    background-color: #3A3845;
+    display:flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    padding-top:10px;
+}
+
+.socialMedia svg{
+    color: white;
+    margin: 15px;
+    font-size:40px;
+    cursor:pointer;
+}
+
+.footer p{
+    color:white;
+}
