+++ src/styles/Home.css
@@ -1,0 +1,67 @@
+.home{
+    width:92%;
+    height:90vh;
+    display:flex;
+    padding-top: 100px;
+    margin:0% 4% 4% 4%;
+}
+
+.leftSide{
+    width:auto;
+    height:100%;
+    flex:50%;
+    display:flex;
+    justify-content: center;
+    align-items: left;
+    flex-direction: column;
+    background-position: center;
+    background-repeat: no-repeat;
+    background-size: cover;
+    margin-left:50px;
+}
+
+.rightSide{
+    height:100%;
+    flex:50%;
+    background-position:center;
+    background-repeat:no-repeat;
+    background-size:cover;
+    font-family:"Gill Sans","Gill Sans MT",Calibri,"Trebuchet MS",sans-serif;
+}
+
+.leftSide h1{
+    justify-content: center;
+    align-items: center;
+    font-size: 60px;
+    font-weight:bold;
+    color:black;
+}
+
+.leftSide p{
+    padding:0;
+    margin:0;
+    font-size: 20px;
+    font-weight: lighter;
+    color:black;
+}
+
+.leftSide button{
+    color:black;
+    border:none;
+    border: radius 5px;
+    cursor:pointer;
+    justify-content: center;
+    background-color:#121619;
+    height: 50px;
+    width: 180px;
+    color: white;
+    text-decoration: none;
+    font-size:20px;
+    margin-top:40px;
+}
+
+.leftSide button:hover{
+    background: #07090a;
+    cursor:pointer;
+    transition: 0.3s ease-in;
+}
