+++ src/styles/Sign.css
@@ -5,7 +5,8 @@
     justify-content: center;
     align-items: center;
     flex-direction: column;
-    padding-top:20px;
+    padding-top:100px;
+    padding-bottom: 40px;
 }
 
 .SignForm h2{
